add scroller to HourlyCard
