+++ src/components/ui/card-small/CardSmall.module.css
@@ -1,0 +1,4 @@
+.list {
+  display: flex;
+  overflow-y: auto;
+}
