+++ src/components/ui/card-small/index.ts
@@ -1,0 +1,1 @@
+export { CardSmall } from "./CardSmall";
