+++ src/utils/buildClassName.ts
@@ -1,0 +1,5 @@
+type Parts = (string | false | undefined)[];
+
+export function buildClassName(...parts: Parts): string {
+  return parts.filter(Boolean).join(" ");
+}
